+++ multisrc/overrides/madara/neatmanga/src/NeatManga.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.neatmanga
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class NeatManga : Madara("NeatManga", "https://neatmanga.com", "en", SimpleDateFormat("dd MMM yyyy", Locale.US))
