+++ multisrc/overrides/madara/nekobreaker/src/NekoBreaker.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.pt.nekobreaker
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class NekoBreaker : Madara("NekoBreaker", "https://nekobreaker.com", "pt-BR", SimpleDateFormat("MMMM dd, yyyy", Locale("pt")))
