+++ multisrc/overrides/madara/nijitranslations/src/NijiTranslations.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.ar.nijitranslations
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class NijiTranslations : Madara("Niji Translations", "https://niji-translations.com", "ar", SimpleDateFormat("MMMM dd, yyyy", Locale("ar")))
