+++ multisrc/overrides/madara/projetoscanlator/src/ProjetoScanlator.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.pt.projetoscanlator
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class ProjetoScanlator : Madara("Projeto Scanlator", "https://projetoscanlator.com", "pt-BR", SimpleDateFormat("dd/MM/yyyy", Locale("pt", "BR")))
