+++ multisrc/overrides/madara/queensmanga/src/QueensManga.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.ar.queensmanga
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+
+class QueensManga : Madara("QueensManga ملكات المانجا", "https://queensmanga.com", "ar") {
+    override fun chapterListSelector(): String = "div.listing-chapters_wrap a"
+}
