+++ multisrc/overrides/madara/randomtranslations/src/RandomTranslations.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.randomtranslations
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class RandomTranslations : Madara("Random Translations", "https://randomtranslations.com", "en", SimpleDateFormat("dd/MM/yyyy", Locale.US))
