+++ multisrc/overrides/madara/readmanhua/src/ReadManhua.kt
@@ -1,0 +1,12 @@
+package eu.kanade.tachiyomi.extension.en.readmanhua
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class ReadManhua : Madara(
+    "ReadManhua",
+    "https://readmanhua.net",
+    "en",
+    dateFormat = SimpleDateFormat("dd MMM yy", Locale.US)
+)
