+++ multisrc/overrides/madara/renascans/src/RenaScans.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.renascans
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class RenaScans : Madara("Renascence Scans (Renascans)", "https://new.renascans.com", "en", SimpleDateFormat("dd/MM/yyyy", Locale.US))
