+++ multisrc/overrides/madara/sixiangscans/src/SiXiangScans.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.sixiangscans
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class SiXiangScans : Madara("SiXiang Scans", "http://www.sixiangscans.com", "en", SimpleDateFormat("MMMM dd, yyyy", Locale.US))
