+++ multisrc/overrides/madara/siyahmelek/src/Siyahmelek.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.tr.siyahmelek
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class Siyahmelek : Madara("Siyahmelek", "https://siyahmelek.com", "tr", SimpleDateFormat("dd MMM yyyy", Locale("tr")))
