+++ multisrc/overrides/madara/soloscanlation/src/SoloScanlation.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.soloscanlation
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class SoloScanlation : Madara("SoloScanlation", "https://soloscanlation.site", "en", SimpleDateFormat("MMMM dd, yyyy", Locale.US))
