+++ multisrc/overrides/madara/spookyscanlations/src/SpookyScanlations.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.es.spookyscanlations
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class SpookyScanlations : Madara("Spooky Scanlations", "https://spookyscanlations.xyz", "es", SimpleDateFormat("MMMM dd, yyyy", Locale("es")))
