+++ multisrc/overrides/madara/threesixtyfivemanga/src/ThreeSixtyFiveManga.kt
@@ -1,0 +1,10 @@
+package eu.kanade.tachiyomi.extension.en.threesixtyfivemanga
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import eu.kanade.tachiyomi.network.GET
+import okhttp3.Request
+
+class ThreeSixtyFiveManga : Madara("365Manga", "https://365manga.com", "en") {
+    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/manga/page/$page/?m_orderby=views", headers)
+    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/manga/page/$page/?m_orderby=latest", headers)
+}
