+++ multisrc/overrides/madara/toongod/src/ToonGod.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.toongod
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class ToonGod : Madara("ToonGod", "https://www.toongod.com", "en", SimpleDateFormat("dd MMM yyyy", Locale.US))
