+++ multisrc/overrides/madara/toonpoint/src/ToonPoint.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.toonpoint
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+
+class ToonPoint : Madara("ToonPoint", "https://toonpoint.com", "en") {
+    override val userAgentRandomizer = ""
+}
