+++ multisrc/overrides/madara/uyuyanbalik/src/UyuyanBalik.kt
@@ -1,0 +1,8 @@
+package eu.kanade.tachiyomi.extension.tr.uyuyanbalik
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.*
+
+class UyuyanBalik : Madara("Uyuyan Balik", "https://uyuyanbalik.com/", "tr", SimpleDateFormat("dd MMMM yyyy", Locale.US))
+
