+++ multisrc/overrides/madara/warqueenscan/src/WarQueenScan.kt
@@ -1,0 +1,8 @@
+package eu.kanade.tachiyomi.extension.pt.warqueenscan
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.*
+
+class WarQueenScan : Madara("War Queen Scan", "https://wqscan.com", "pt-BR", SimpleDateFormat("dd/MM/yyyy", Locale.US))
+
