+++ multisrc/overrides/madara/webtoonxyz/src/WebtoonXYZ.kt
@@ -1,0 +1,11 @@
+package eu.kanade.tachiyomi.extension.en.webtoonxyz
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import eu.kanade.tachiyomi.network.GET
+import okhttp3.Request
+
+class WebtoonXYZ : Madara("WebtoonXYZ", "https://www.webtoon.xyz", "en") {
+    private fun pagePath(page: Int) = if (page > 1) "page/$page/" else ""
+    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/webtoons/${pagePath(page)}?m_orderby=views", headers)
+    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/webtoons/${pagePath(page)}?m_orderby=latest", headers)
+}
