+++ multisrc/overrides/madara/worldromancetranslation/src/WorldRomanceTranslation.kt
@@ -1,0 +1,8 @@
+package eu.kanade.tachiyomi.extension.id.worldromancetranslation
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.*
+
+class WorldRomanceTranslation : Madara("World Romance Translation", "https://wrt.my.id/", "id", SimpleDateFormat("dd MMMM yyyy", Locale("id")))
+
