+++ multisrc/overrides/madara/wuxiaworld/src/WuxiaWorld.kt
@@ -1,0 +1,13 @@
+package eu.kanade.tachiyomi.extension.en.wuxiaworld
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import eu.kanade.tachiyomi.network.GET
+import eu.kanade.tachiyomi.source.model.FilterList
+import okhttp3.Request
+
+class WuxiaWorld : Madara("WuxiaWorld", "https://wuxiaworld.site", "en") {
+    override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/tag/webcomic/page/$page/?m_orderby=views", headers)
+    override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/tag/webcomic/page/$page/?m_orderby=latest", headers)
+    override fun searchMangaRequest(page: Int, query: String, filters: FilterList) = super.searchMangaRequest(page, "$query comics", filters)
+    override fun popularMangaNextPageSelector() = "div.nav-previous.float-left"
+}
