+++ multisrc/overrides/madara/zinmanga/src/ZinManga.kt
@@ -1,0 +1,9 @@
+package eu.kanade.tachiyomi.extension.en.zinmanga
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import okhttp3.Headers
+
+class ZinManga : Madara("Zin Translator", "https://zinmanga.com", "en") {
+    override fun headersBuilder(): Headers.Builder = super.headersBuilder()
+        .add("Referer", "https://zinmanga.com/")
+}
