+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/madara/MadaraGenerator.kt
@@ -80,6 +80,71 @@ class MadaraGenerator : ThemeSourceGenerator {
             SingleLang("Kombatch", "https://kombatch.com", "id"),
 
             SingleLang("Mangazuki.online", "http://mangazukinew.online", "en", className = "MangazukiOnline"),
+
+            SingleLang("Nazarick Scans", "https://nazarickscans.com", "en"),
+            SingleLang("NeatManga", "https://neatmanga.com", "en"),
+            SingleLang("NekoBreaker", "https://nekobreaker.com", "pt-BR"),
+            SingleLang("NekoScan", "https://nekoscan.com", "en"),
+            SingleLang("Neox Scanlator", "https://neoxscans.com", "pt-BR"),
+            SingleLang("Night Comic", "https://www.nightcomic.com", "en"),
+            SingleLang("Niji Translations", "https://niji-translations.com", "ar"),
+            SingleLang("Ninjavi", "https://ninjavi.com", "ar"),
+            SingleLang("NTS Void Scans", "https://ntsvoidscans.com", "en"),
+            SingleLang( "Off Scan", "https://offscan.top", "pt-BR"),
+            SingleLang("مانجا اولاو", "https://olaoe.giize.com", "ar", className = "OlaoeManga"),
+            SingleLang("OnManga", "https://onmanga.com", "en"),
+            SingleLang("Origami Orpheans", "https://origami-orpheans.com.br", "pt-BR"),
+            SingleLang("PMScans", "https://www.pmscans.com", "en"),
+            SingleLang("Pojok Manga", "https://pojokmanga.com", "id"),
+            SingleLang("PornComix", "https://www.porncomixonline.net", "en", isNsfw = true),
+            SingleLang("Prime Manga", "https://primemanga.com", "en"),
+            SingleLang("Projeto Scanlator", "https://projetoscanlator.com", "pt-BR"),
+            SingleLang("QueensManga ملكات المانجا", "https://queensmanga.com", "ar", className = "QueensManga"),
+            SingleLang("Raider Scans", "https://raiderscans.com", "en"),
+            SingleLang("Random Translations", "https://randomtranslations.com", "en"),
+            SingleLang("Raw Mangas", "https://rawmangas.net", "ja", isNsfw = true),
+            SingleLang("ReadManhua", "https://readmanhua.net", "en"),
+            SingleLang("Renascence Scans (Renascans)", "https://new.renascans.com", "en", className = "RenaScans"),
+            SingleLang("Rüya Manga", "https://www.ruyamanga.com", "tr", className = "RuyaManga"),
+            SingleLang("S2Manga", "https://s2manga.com", "en"),
+            SingleLang("Sekte Doujin", "https://sektedoujin.xyz", "id", isNsfw = true),
+            SingleLang("ShoujoHearts", "http://shoujohearts.com", "en"),
+            SingleLang("SiXiang Scans", "http://www.sixiangscans.com", "en"),
+            SingleLang("Siyahmelek", "https://siyahmelek.com", "tr", isNsfw = true),
+            SingleLang("Skymanga", "https://skymanga.co", "en"),
+            SingleLang("SoloScanlation", "https://soloscanlation.site", "en"),
+            SingleLang("Spooky Scanlations", "https://spookyscanlations.xyz", "es"),
+            SingleLang("StageComics", "https://stagecomics.com", "pt-BR"),
+            SingleLang("TheTopComic", "https://thetopcomic.com", "en"),
+            SingleLang("365Manga", "https://365manga.com", "en", className = "ThreeSixtyFiveManga"),
+            SingleLang("ToonGod", "https://www.toongod.com", "en"),
+            SingleLang("Toonily", "https://toonily.com", "en", isNsfw = true),
+            SingleLang("Toonily.net", "https://toonily.net", "en", isNsfw = true),
+            SingleLang("ToonPoint", "https://toonpoint.com", "en"),
+            SingleLang("Top Manhua", "https://topmanhua.com", "en"),
+            SingleLang("TritiniaScans", "https://tritinia.com", "en"),
+            SingleLang("TruyenTranhAudio.com", "https://truyentranhaudio.com", "vi"),
+            SingleLang("TruyenTranhAudio.online", "https://truyentranhaudio.online", "vi", className = "truyentranhaudioonline"),
+            SingleLang("Tsubaki No Scan", "https://tsubakinoscan.com", "fr"),
+            SingleLang("Türkçe Manga", "https://turkcemanga.com", "tr"),
+            SingleLang("Twilight Scans", "https://twilightscans.com", "en"),
+            SingleLang("Uyuyan Balik", "https://uyuyanbalik.com/", "tr"),
+            SingleLang("Vanguard Bun", "https://vanguardbun.com/", "en"),
+            SingleLang("Void Scans", "https://voidscans.com", "en"),
+            SingleLang("Wakascan", "https://wakascan.com", "fr"),
+            SingleLang("War Queen Scan", "https://wqscan.com", "pt-BR"),
+            SingleLang("WebNovel", "https://webnovel.live", "en"),
+            SingleLang("WebToonily", "https://webtoonily.com", "en"),
+            SingleLang("WebtoonXYZ", "https://www.webtoon.xyz", "en"),
+            SingleLang("WeScans", "https://wescans.xyz", "en"),
+            SingleLang("WoopRead", "https://woopread.com", "en"),
+            SingleLang("World Romance Translation", "https://wrt.my.id/", "id"),
+            SingleLang("WuxiaWorld", "https://wuxiaworld.site", "en"),
+            SingleLang("Yaoi Toshokan", "https://yaoitoshokan.com.br", "pt-BR", isNsfw = true),
+            SingleLang("Yokai Jump", "https://yokaijump.fr", "fr"),
+            SingleLang("Yuri Verso", "https://yuri.live", "pt-BR"),
+            SingleLang("Zin Translator", "https://zinmanga.com", "en"),
+            SingleLang("ZManga", "https://zmanga.org", "es"),
     )
 
     companion object {
